bottom action bar and favortie fragment

+++ app/src/main/java/com/project/weather/model/Hourly.kt
@@ -22,5 +22,5 @@ data class Hourly(
     @SerializedName("wind_gust")
     val windGust: Double,
     val weather: List<Weather>,
-    val pop: Long
+    val pop: Double
 ) : Serializable
